+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 * Add json schema and wrapper for imaging data ``life/qbic/datamodel/datasets/imaging/ImageMetadata.groovy``
 
+* Provide new classes that describe incoming nf-core bioinformatic pipeline data structures and contains it in a dedicated NfCorePipelineResult class (`#198 <https://github.com/qbicsoftware/data-model-lib/pull/198>`_)
+
 **Fixed**
 
 **Dependencies**